requirements.txt
Django==4.2.4
celery==5.3.1 
psycopg2==2.9.6
python-dotenv==1.0.0
django-background-tasks==1.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-channels==0.7.0
django-compat==1.0.15
django-cookies-samesite==0.9.0
django-cors-headers==3.13.0
django-cron==0.6.0
django-inspectdb-refactor==0.4
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytest==7.4.0
pytest-django==4.5.2
pytest-mock==3.11.1
eventlet==0.33.3
stripe==6.2.0
kombu==5.3.1